// apps/dashboard/src/payload-generated-schema.ts
export const enum_organization_members_role = pgEnum('enum_organization_members_role', ['owner', 'admin', 'member']);
// ...
export const users = pgTable(
  'users',
  {
    id: varchar('id').primaryKey(),
    name: varchar('name'),
    telegramUsername: varchar('telegram_username'),
    discordUsername: varchar('discord_username'),
    twoFactorEnabled: boolean('two_factor_enabled').default(false),
    twoFactorSecret: varchar('two_factor_secret'),
    backupCode: varchar('backup_code'),
    updatedAt: timestamp('updated_at', { mode: 'string', withTimezone: true, precision: 3 }).defaultNow().notNull(),
    createdAt: timestamp('created_at', { mode: 'string', withTimezone: true, precision: 3 }).defaultNow().notNull(),
    email: varchar('email').notNull(),
    resetPasswordToken: varchar('reset_password_token'),
    resetPasswordExpiration: timestamp('reset_password_expiration', {
      mode: 'string',
      withTimezone: true,
      precision: 3,
    }),
    salt: varchar('salt'),
    hash: varchar('hash'),
    loginAttempts: numeric('login_attempts', { mode: 'number' }).default(0),
    lockUntil: timestamp('lock_until', { mode: 'string', withTimezone: true, precision: 3 }),
  },
  (columns) => [
    index('users_updated_at_idx').on(columns.updatedAt),
    index('users_created_at_idx').on(columns.createdAt),
    uniqueIndex('users_email_idx').on(columns.email),
  ],
);
// ...
export const organizations = pgTable(
  'organizations',
  {
    id: varchar('id').primaryKey(),
    name: varchar('name').notNull(),
    slug: varchar('slug'),
    createdBy: varchar('created_by').notNull(),
    quotaBalance: numeric('quota_balance', { mode: 'number' }).notNull().default(100000000),
    rpsLimit: numeric('rps_limit', { mode: 'number' }).notNull().default(100000),
    rpsPaidLimit: numeric('rps_paid_limit', { mode: 'number' }).notNull().default(100000),
    rpsExpiresAt: timestamp('rps_expires_at', { mode: 'string', withTimezone: true, precision: 3 }),
    quotaUsed: numeric('quota_used', { mode: 'number' }).notNull().default(0),
    rpsForever: boolean('rps_forever').default(true),
    updatedAt: timestamp('updated_at', { mode: 'string', withTimezone: true, precision: 3 }).defaultNow().notNull(),
    createdAt: timestamp('created_at', { mode: 'string', withTimezone: true, precision: 3 }).defaultNow().notNull(),
  },
  (columns) => [
    uniqueIndex('organizations_slug_idx').on(columns.slug),
    index('organizations_created_by_idx').on(columns.createdBy),
    index('organizations_updated_at_idx').on(columns.updatedAt),
    index('organizations_created_at_idx').on(columns.createdAt),
  ],
);

export const organization_members = pgTable(
  'organization_members',
  {
    id: varchar('id').primaryKey(),
    organization: varchar('organization_id')
      .notNull()
      .references(() => organizations.id, {
        onDelete: 'set null',
      }),
    user: varchar('user_id')
      .notNull()
      .references(() => users.id, {
        onDelete: 'set null',
      }),
    role: enum_organization_members_role('role').notNull().default('member'),
    updatedAt: timestamp('updated_at', { mode: 'string', withTimezone: true, precision: 3 }).defaultNow().notNull(),
    createdAt: timestamp('created_at', { mode: 'string', withTimezone: true, precision: 3 }).defaultNow().notNull(),
  },
  (columns) => [
    index('organization_members_organization_idx').on(columns.organization),
    index('organization_members_user_idx').on(columns.user),
    index('organization_members_updated_at_idx').on(columns.updatedAt),
    index('organization_members_created_at_idx').on(columns.createdAt),
    uniqueIndex('organization_user_idx').on(columns.organization, columns.user),
  ],
);